reviewed branch #36
Everything worked except in method for adding a auction to favourites
list you could add the same auction multiple times. This was basically
the same problem we had in the delete function so I added the same
logic.
In the delete method I deleted the line that check if the auction exist
in the auction db, and thats because lets say an auction is deleted then
you cant delete it from your favourites list if it happends to be there.